requirements.txt
mysql-connector-python
python-dotenv
Flask-SQLAlchemy
Flask
pymysql